refactor(process-list): split modules
